begin work on a reporting DSL
So our reports basically are columns, and rows (data tuple for each
column). This commit creats a very small DSL to define columns that
should be available in a report.

Usage:

class ConsolidatedReport < Report
  include Mostfit::Reporting

  column :loan_amount => [:applied, :sanctioned, :disbursed]
  column :balance_outstanding => [:principal, :interest, :total]
end

This doesn't do much now, except that you can say
ConsolidatedReport.columns to get columns that should be available for
ConsolidatedReports.