Fix license text.
LICENSE.txt mostly contained the text of the GNU AGPL v3, but there were
changes:
* Some errors existed. Some characters were missing, and a line was
  deleted.
* Summary text and an explanation were present at the beginning of the
  license.

Obviously the errors needed fixing, but the summary text should be
removed, too. It is better to use the license text as-is, as provided by
the FSF (Free Software Foundation). Any modifications may inadvertently
alter the license, making the software incompatible with other AGPL
software.

The FSF also recommends keeping the license text in a file named
COPYING. Following this convention should make it even easier for folks
to discover and follow the Mostfit licensing terms.