re-write parts of rule's ui logic
This fixes StaffMember-Manager problem we had in the rules UI. Earlier,
on selecting a manager, the next select list would only contain a list
of names of all managers. Selecting manager now, brings up a list of all
properties, and relations of manager (a.k.a staff member) to choose
from.

This should take care of other similar cases too, and hopefully closes
issue #680.

Thanks to Prashant Tibrewal for helping out with this one.