Feature 1058: Capability for transaction log to include extra information in a generic fashion.
Each kind of transaction may include additional information that is specific to the nature of the transaction.
For example, loan transactions may add information about the loan product, funder, funding line, etc.
The nature of such information may vary based on different kinds of transactions that are supported in Mostfit at different points in time.

Such information may be captured as extended information in a generic manner as follows: each form of extended information will be assigned a type (such as Loan Product, Funding Line, etc.). The type, the identifier, and the value for each such item for extended information are the only information captured. Transaction log entries will have one or more extended information items

The allowed types can be defined in constants

To obtain information about extended_info_items, each transaction that is observed should implement a method such as extended_info that returns an array of hashes, each hash representing one extended_info_item. The transaction observer merely calls this method and uses the information that is returned.