Feature #1074: Introduced organisation and domain information in transaction and event logs
Each transaction log entry specifies an organization and a domain
In the case of transaction log entries, information on organisation (and any domain) come
from the transaction itself.
The effective date of the transaction is used to determine the corresponding organisation.
The domain, if any, is specified in the transaction itself.
The organisation (and domain, if any) are specified only using the GUID for the same