Feature #1061: A rake task to export transaction logs and event logs
The rake task accepts argumetns that are begin date and optionally and end date.
If the end date is now supplied, all entried till date are exported.

For transaction logs the date range used to select the log entries are
based on the effective date of the log entry

For event logs the date range used to select the log entries are based
 on the date of creation of the log entry

The rake task creates an xml file with the name in the following format:-
<log_type_name>.<org_guid>.from.<from_date>.to.<to_date>.xml