Bug #1215: The accounting action for a change of :create  in event log is now :create and not :allow_posting
Whenever the create method is observed by the model observer
on whatever models it is observing, the event accounting
action for that is :create.

This works fine when a Client is created.

But in some cases, like when we assign a loan to a client
(i.e.) when we are creating a loan, the update function is
observed to be called and the create function is not observed
to be called by the Datamapper::Observer that we are using
in the Model Observer.