This commit is with a view to providing all users with an interactive shell for Mostfit. The following changes have been implemented
DataAccessObserver now observes all models.
When called through a controller, it gets the session variables from the session so it can find the current user.

To be used from a shell, you will have to set ENV['MOSTFIT_USER'] and ENV['MOSTFIT_PASSWORD'].
These will allow the DAO to get the correct authorisations for the given user.
Now all data access, even from a shell is logged.
Even the destroy! methods are logged and disallowed for anyone other than admin.
For any changes, it then runs @_user.is_manager_of?(obj) and raises NotPrivileged if it returns false.

It is my belief that these hooks cannot be overridden using fancy ruby metaprogramming because they are implemented at server boot time.
The only way to override them is to change the source code on file.

This means that for users whose ssh keys are on the mostfit server, they can now do something like this
$ ssh mostfit.in -t 'cd <mostfit_dir>; merb -i'

and get a fully authorised console.

Bye bye pesky delete this and that in bulk requests!