new-clients functionality merged into takeover
See the diff of app/controllers/clients.rb for clues. Basically you can now add as many subclasses to Client as you want
and if you want to massage the params for the new subclass, you can do so in the controller.

Also you can add views in app/views/clients/<new_model_name> and they will be used instead of the default ones.