Merged new holidays functionality into new-layout branch.
Holidays is working but the code is fugly. For some reason the has n, :holidays, :through => Resource
never worked for me and now we have the super ugly holidays_for model with all the crappy pluralisation
that comes with it.

Also, there is no front end for holidays but the backend seems to be working fine. No specs of course....

To sum up, we now have three models Holiday, HolidayCalendar and HolidaysFor (the join model)

To add a holiday, do something like hc = HolidayCalendar.first; hc.add_holiday(<holiday or holioday_id>);

This will add all the affected loans to DirtyLoans and start a new thread to work itself through updating their history.

Conflicts:
	lib/irb.rb