Bug #1520: Function for displaying a journal type
A function has been defined in the Journal type class
for displaying the Journal type with a specific color
in html pages. This function imitates the logic found
in the page listing accounting rules where Journal
type has been displayed with a specific color background.

The order of the columns of Credit and Debit have
been interchanged in the page listing the accounting
rules. The displaying of JournalType has been replaced
with the more generic function that has been written.

In the general listing of journals the above mentioned
function has been used.