Bug #1520: Function for displaying an account name
When displaying the list of journals, when the account
name is specified it does not give much information as
accounts having the same name can belong to different
branches. Hence, a function display_name has been added
to the Account model for displaying the account name and
its corresponding branch name in brackets to the RHS of
the account name. The corresponding change has also been
made to the journal listing page as well.