Bug #1519: Disallow deletion of vouchers(journals) and permit only reversal
The link for deletion of vouchers has been removed from
the user screen but the deletion of vouchers has not
been disallowed in the system per se. One can still delete
journals from the irb.

Reversing a journal is the entry of a new journal that
is the same in all respects, except that the postings
have the opposite effect on accounts as that of the
original journal.

The narration in the reversing journal uses the narration
of the original journal with a prefix: "REVERSED:" added
as a hint to the user.

              !!! WARNING !!!
A journal once reversed should not be allowed to be
reversed again. But in the current case that is possible.
One can reverse an already reversed journal as many times
as one pleases. A solution has to be discussed on this
matter. One possible way to achieve this is by adding
a Boolean field in the Journal model called reversed
which will be set when a journal is reversed and will
be reset otherwise.