fixed a bug in cachers that saw us lose total data integrity, silently. Sometimes, perhaps due to not wrapping the whole thing in a transaction, there would appear two rows for a center for the same date. This is of course, double counting and totally unacceptable.
I have fixed it by doing what LoanHistory does. I dropped the id column and added keys to some fields so that we are guaranteed uniqueness
at a database level and we don't have to ever worry about race conditions and transactions and locks.

I also wrapped BranchCache.update in a transaction, so if it crashes, the db will be rolled back.

There are some additions in lib/tasks/prepare_db.rake, so running "rake mostfit:db:prepare" should be sufficient.

BE WARNED - this will drop and recreate all the cachers so will take a long time if you have a lot of them